pyproject.toml
[project]
name = "hotel-reservation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-storage>=3.5.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "joblib>=1.5.2",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "setuptools>=80.9.0",
    "tqdm>=4.67.1",
]
